requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.1.12
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
djoser==2.2.0
dnspython==2.4.2
idna==3.4
mongoengine==0.27.0
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
pymongo==3.12.3
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.2.4
typing_extensions==4.8.0
urllib3==2.0.6
